Integrate callgraph library into the main CMake
We used to assume that the callgraph library would be installed on the
user's system, but that is a bit intrusive. Instead, we now integrate it
into the main CMake flow and pass the appropriate paths around.

We also rename the files from 'cg.*' to 'callgraph.*', since we're no
longer installing to a 'callgraph' directory.